Remove the requirement to download artifacts
Before this, ffmpeg and EasyOCR needed some initialization code that needed to run exactly once. Since We now expose a python API, that needs to work without the user having to call some magic setup function.
The function for preparing all downloadable artifacts is still there, so we can prepare the docker image to run in isolated environments.